user activity

+++ PSTRiverPlugin/src/main/java/com/fruitfactory/pstriver/useractivity/IInputHookIdle.java
@@ -10,5 +10,5 @@ package com.fruitfactory.pstriver.useractivity;
  * @author Yariki
  */
 public interface IInputHookIdle {
-    int getIdleTime();
+    long getIdleTime();
 }
